move rest api tests and code to feature folder (#36044)
